move routes into a routes module
