+++ app/guest/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 from flask import Blueprint
-from . import add_guest
+from . import routes
 
 blueprint = Blueprint('guest', __name__)
-add_guest.routes(blueprint)
+routes.register_routes(blueprint)
